librmf: report error string on error messages
This is most usually filled in with the server-side GError message, with
the exception of the method to start a connection on a network, where the
call end reason string is reported back to the client. On the client side,
this is the only situation too where the string is extracted and shown
on exceptions.