rmfd: assume unlocked only whenever SIM is fully ready
After an unlock operation, there may be some time where the modem reports that
SIM-PIN is unlocked, but still the SIM is not fully initialized, see the
following logs:

Executing the following:
    while [ 1 ]; do
        echo "----------------------------"
        sudo rmfcli --is-locked
        sudo rmfcli --get-imsi
        usleep 1000
    done

We get the following output:
    ----------------------------
    PIN is locked
    Exception: UIM uninitialized
    ----------------------------
    PIN is locked
    Exception: UIM uninitialized
    ----------------------------
    PIN is locked
    Exception: UIM uninitialized

Here PIN is sent

    ----------------------------
    PIN is unlocked                     < WRONG!
    Exception: UIM uninitialized
    ----------------------------
    PIN is unlocked                     < WRONG!
    Exception: UIM uninitialized
    ----------------------------
    PIN is unlocked                     < OK
    IMSI: 262014242065367
    ----------------------------
    PIN is unlocked                     < OK
    IMSI: 262014242065367

So, update rmfd so that the 'Is Unlocked' operation only returns that the SIM is
unlocked when it has been fully initialized. With the change, the same execution
gives us:

    ----------------------------
    PIN is locked
    Exception: UIM uninitialized
    ----------------------------
    PIN is locked
    Exception: UIM uninitialized
    ----------------------------
    PIN is locked
    Exception: UIM uninitialized

Here PIN is sent

    ----------------------------
    PIN is unlocked                     < OK
    IMSI: 262014242065367
    ----------------------------
    PIN is unlocked                     < OK
    IMSI: 262014242065367

Along with this change, the libqmi requirement was updated to 1.9 (development
version from git), in order to use the new "UIM Get Card Status" message
support.