rmfd,stats: avoid notifying partial record multiple times
If a partial record is found as the last record we'll be notifying in syslog
about its contents. But once that is done, we'll flag the record as Final so
that the syslog emission is not done when rmfd restarts.

With this logic on, partial records can only happen at the end of the stats
file, as every other end-of-connection partial record that happened before would
have already been fixed as final.