test: power status commands at the very end
My test operator takes a long time to register in the network, because it's a
MNVO without its own network, and therefore it needs to poke to the available
networks around to check which one is the allowed one.

So, avoid setting Low power state before the connection attempt; as that would
end up requesting a re-registration in the network from scratch.

Also, add the new power cycle command as the very last command to test.