librmf: change recv() logic to avoid blocking situations
We were doing a blocking recv() to read RMF_MESSAGE_MAX_SIZE bytes,
assuming that the server would close the connection when reply fully
written, even if less than RMF_MESSAGE_MAX_SIZE. In this case, we would
correctly finish the operation only if the full message was received
after the server closed the socket. If the socket was closed before the
full message was received, the logic would end up doing an infinite loop
of recv() calls returning 0... :/

Instead of that, we now read the first 4 bytes of the message, which
include the full message size, and then request to read the whole full
message, not a single byte more. If any of the read calls doesn't match
the size we expect to read, we'll return an error.

Also, we now provide an explicit timeout for each recv() operation, so
that it doesn't block forever in any case.